# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateUserRequest:
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        maxLength: 100
        minLength: 10
        type: string
    required:
    - email
    - name
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: user not found
        type: string
    type: object
  dto.ErrorResponseServer:
    properties:
      error:
        example: Internal server error
        type: string
    type: object
  dto.ErrorResponseValidation:
    properties:
      error:
        example: Validation error
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        example: User deleted successfully
        type: string
    type: object
...
  dto.UpdateUserRequest:
    properties:
      age:
        minimum: 0
        type: integer
      city:
        maxLength: 255
        type: string
      country:
        type: string
      email:
        type: string
      employment:
        type: boolean
      gender:
        enum:
        - male
        - female
        type: string
      image_url:
        type: string
      migration:
        type: boolean
      name:
        maxLength: 100
        minLength: 10
        type: string
      phone:
        type: string
      salary:
        type: number
      state:
        maxLength: 255
        type: string
    type: object
  dto.UserResponse:
    properties:
      admin:
        type: boolean
      age:
        type: integer
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      employment:
        type: boolean
      gender:
        type: string
      id:
        type: string
      image_url:
        type: string
      migration:
        type: boolean
      name:
        type: string
      phone:
        type: string
      salary:
        type: number
      state:
        type: string
      updated_at:
        type: string
    type: object
  dto.UsersResponse:
    properties:
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: dafondeveloper@gmail.com
    name: DafonCV Support
    url: https://www.dafoncv.online
  description: API for consultancy business and cv generation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Dafon CV API
  version: "1.0"
paths:
...
  /api/v1/user:
    post:
      consumes:
      - application/json
      description: Creates a new user
      parameters:
      - description: Create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      summary: Create user
      tags:
      - user
  /api/v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Updates an existing user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid user ID format or validation error
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Update user by ID
      tags:
      - user
  /api/v1/user/all:
    get:
      consumes:
      - application/json
      description: Returns a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UsersResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponseValidation'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponseServer'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  UserIDHeader:
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"